Image Carousel
As well as updating the social media bits more, fixing a naming, and
tinkering.
----
If you load up the main page, look  If you look at the left arrow's
bottom, and then at the right arrow, you'll see the 'missing part'
issue.
Badger: "Yeah, it's your css.  Your css says that any div anywhere
inside of a span9 should have a radius and background. Which overwrites
the buttons.  Ideally, the "make this div have background and rounded
edges", should be a class. On the div. Not every single div underneath
.span9."
Badger: "  .container-1 .row-fluid .span9 div { <<-- That's the
offending rule."