Update code; is now Production-ready!
To run in Production mode:
./grunt build && PORT=9001 ./grunt serve:dist

Get rid of some assets which are no longer necessary.
Update upload code to find assets on filesystem and update
database links accordingly. This is so as to perform automatic
file-discovery when minified/rev'd images are used in Production.

You know, so images actually work in production when they are
name-mangled by the dist (grunt build) tasks.